+++ packages/app/components/loading.component.tsx
@@ -1,0 +1,20 @@
+import React from 'react'
+import { ActivityIndicator, View, StyleSheet } from 'react-native'
+
+export const LoadingComponent = () => (
+  <View style={[styles.container, styles.horizontal]}>
+    <ActivityIndicator size="large" />
+  </View>
+)
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+  },
+  horizontal: {
+    flexDirection: 'row',
+    justifyContent: 'space-around',
+    padding: 10,
+  },
+})