+++ packages/app/components/login.component.tsx
@@ -21,6 +21,7 @@ import ActionSheet from 'rn-actionsheet-module'
 import { useAsyncStorage } from 'use-async-storage'
 import { schema } from '../app.json'
 import { Colors, Layout, Sizing } from '../styles'
+import { translate } from '../utils/translation'
 import {
   CloseOutlineIcon,
   PersonIcon,
@@ -43,14 +44,16 @@ export const Login = () => {
     'loginMethodIndex',
     '0'
   )
+
   const loginMethods = [
-    'Öppna BankID på denna enhet',
-    'Öppna BankID på annan enhet',
-    'Logga in som testanvändare',
+    translate('auth.bankid.OpenOnThisDevice'),
+    translate('auth.bankid.OpenOnAnotherDevice'),
+    translate('auth.loginAsTestUser'),
   ]
+
   const selectLoginMethod = () => {
     const options = {
-      title: 'Välj inloggningsmetod',
+      title: translate('auth.chooseLoginMethod'),
       optionsIOS: loginMethods,
       optionsAndroid: loginMethods,
       onCancelAndroidIndex: loginMethodIndex,
@@ -110,7 +113,7 @@ export const Login = () => {
           : `bankid:///?autostarttoken=${token}&redirect=null`
       Linking.openURL(bankIdUrl)
     } catch (err) {
-      setError('Öppna BankID manuellt')
+      setError(translate('auth.bankid.OpenManually'))
     }
   }
 
@@ -133,7 +136,7 @@ export const Login = () => {
       status.on('PENDING', () => console.log('BankID app not yet opened'))
       status.on('USER_SIGN', () => console.log('BankID app is open'))
       status.on('ERROR', () => {
-        setError('Inloggningen misslyckades, försök igen!')
+        setError(translate('auth.loginFailed'))
         showModal(false)
       })
       status.on('OK', () => console.log('BankID ok'))
@@ -148,7 +151,7 @@ export const Login = () => {
       <View style={styles.loginForm}>
         {loginMethodIndex === 1 && (
           <Input
-            label="Personnummer"
+            label={translate('general.socialSecurityNumber')}
             autoFocus
             value={socialSecurityNumber}
             style={styles.pnrInput}
@@ -162,7 +165,7 @@ export const Login = () => {
             onSubmitEditing={(event) => startLogin(event.nativeEvent.text)}
             caption={error || ''}
             onChangeText={(text) => handleInput(text)}
-            placeholder="Ditt personnr"
+            placeholder={translate('auth.placeholder_SocialSecurityNumber')}
           />
         )}
         <ButtonGroup style={styles.loginButtonGroup}>
@@ -194,7 +197,9 @@ export const Login = () => {
         onBackdropPress={() => showModal(false)}
       >
         <Card disabled>
-          <Text style={styles.bankIdLoading}>Väntar på BankID...</Text>
+          <Text style={styles.bankIdLoading}>
+            {translate('auth.bankid.Waiting')}
+          </Text>
 
           <Button
             onPress={() => {
@@ -202,7 +207,7 @@ export const Login = () => {
               showModal(false)
             }}
           >
-            Avbryt
+            {translate('general.abort')}
           </Button>
         </Card>
       </Modal>
@@ -218,7 +223,7 @@ const styles = StyleSheet.create({
   loginForm: {
     ...Layout.mainAxis.flexStart,
     ...Layout.crossAxis.flexEnd,
-    paddingHorizontal: Sizing.t5,
+    paddingHorizontal: Sizing.t4,
   },
   pnrInput: { minHeight: 70 },
   loginButtonGroup: {